pyproject.toml
[tool.poetry]
name = "llm-endpoint"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
openai = "1.16.2"
langchain = "^0.1.14"
langchain-google-genai = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
